requirements.txt
fastapi
uvicorn
python-multipart
PyYAML
python-dotenv
pypdf
langchain
openai
tiktoken
pinecone-client
unstructured
pandas
numpy
